requirements.txt
asgiref==3.8.1
Django==4.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
sqlparse==0.5.0
asgiref==3.8.1
psycopg2-binary==2.9.9

